Tidying things up

+++ app.py
@@ -2,12 +2,17 @@ import uuid
 import os
 import json
 from flask import Flask, request, jsonify
-from memcached_lib.memcached_client import cache_file
+from memcached_lib.memcached_client import (
+    cache_file,
+    read_file,
+    FileNotFoundExcpetion
+)
 
 app = Flask(__name__)
 
-@app.route('/', methods=['GET', 'POST'])
-def save():
+@app.route('/cache', methods=['GET', 'POST'])
+def cache():
+    #would probably not put both GET and POST methods in same
     if request.method == 'POST':
         file_id = uuid.uuid4()
         file_name = '%s.dat' % file_id
@@ -16,7 +21,25 @@ def save():
             f.write(json_data['data'])
             f.close()
 
-        cache_file(file_name)
-        os.remove(file_name)
+        try:
+            cache_file(file_name)
+        except FileTooLargeException as e:
+            return jsonify(error='Data too large. Must be < 50MB'), 500
 
-    return jsonify(id=file_id), 200
+        os.remove(file_name)
+        return jsonify(id=file_id), 200
+
+    if request.method == 'GET':
+        id = request.args.get('id')
+
+        file_name = '%s.dat' % id
+
+        try:
+            file = read_file(file_name)
+
+            f = open(file, 'r')
+            output = f.read()
+            os.remove(file_name)
+            return jsonify(id = id, data=output), 200
+        except FileNotFoundExcpetion as e:
+            return jsonify(error='File Not Found in Cache'), 500
